indicator: Set a shorter maximum duration for short indication tones.
When a short indication tone has finished playing, we don't want to keep a
silent silent stream around for 60s. Thus, we use a shorter duration after
which the stream will be closed. This way, we won't keep PulseAudio sinks
unsuspended for extended periods of time just to play silence.

Longer tones are not affected by this patch, so they will still be played for
60s unless stopped before that. Of particular interest is TONE_RADIO_ACK, which
is a short tone in the CEPT/ANSI/AT&T standards but a long one in the Japan
standard. Thus, only the timeout of the CEPT/ANSI/AT&T version was modified.

The shorter timeout was arbitrarily chosen to be 5s. Hopefully this is
an appropriate value.